Consolidate and rename extension method classes
