+++ src/Core/src/Handlers/Editor/EditorHandler.Android.cs
@@ -33,7 +33,7 @@ namespace Microsoft.Maui.Handlers
 			
 		public static void MapPredictiveText(EditorHandler handler, IEditor editor)
 		{
-			handler.TypedNativeView?.UpdatePredictiveText(editor);
+			handler.TypedNativeView?.UpdateIsTextPredictionEnabled(editor);
 		}
 	}
 }
