+++ src/Core/src/Platform/iOS/TextFieldExtensions.cs
@@ -4,22 +4,89 @@ namespace Microsoft.Maui
 {
 	public static class TextFieldExtensions
 	{
-		public static void UpdateCharacterSpacing(this UITextField textField, ISearchBar searchBar)
+		public static void UpdateText(this UITextField textField, IEntry entry)
 		{
-			var textAttr = textField.AttributedText?.AddCharacterSpacing(searchBar.Text, searchBar.CharacterSpacing);
+			textField.Text = entry.Text;
+		}
+
+		public static void UpdateTextColor(this UITextField textField, IEntry entry)
+		{
+			textField.UpdateTextColor(entry, null);
+		}
+
+		public static void UpdateTextColor(this UITextField textField, IEntry entry, UIColor? defaultTextColor)
+		{
+			if (entry.TextColor == Color.Default)
+			{
+				if (defaultTextColor != null)
+					textField.TextColor = defaultTextColor;
+			}
+			else
+				textField.TextColor = entry.TextColor.ToNative();
+		}
+
+		public static void UpdateIsPassword(this UITextField textField, IEntry entry)
+		{
+			if (entry.IsPassword && textField.IsFirstResponder)
+			{
+				textField.Enabled = false;
+				textField.SecureTextEntry = true;
+				textField.Enabled = entry.IsEnabled;
+				textField.BecomeFirstResponder();
+			}
+			else
+				textField.SecureTextEntry = entry.IsPassword;
+		}
+
+		public static void UpdateHorizontalTextAlignment(this UITextField textField, ITextAlignment textAlignment)
+		{
+			// We don't have a FlowDirection yet, so there's nothing to pass in here. 
+			// TODO: Update this when FlowDirection is available 
+			// (or update the extension to take an ILabel instead of an alignment and work it out from there) 
+			textField.TextAlignment = textAlignment.HorizontalTextAlignment.ToNative(true);
+		}
+
+		public static void UpdateIsTextPredictionEnabled(this UITextField textField, IEntry entry)
+		{
+			if (entry.IsTextPredictionEnabled)
+				textField.AutocorrectionType = UITextAutocorrectionType.Yes;
+			else
+				textField.AutocorrectionType = UITextAutocorrectionType.No;
+		}
+
+		public static void UpdatePlaceholder(this UITextField textField, IEntry entry)
+		{
+			textField.Placeholder = entry.Placeholder;
+		}
+
+		public static void UpdateIsReadOnly(this UITextField textField, IEntry entry)
+		{
+			textField.UserInteractionEnabled = !entry.IsReadOnly;
+		}
+
+		public static void UpdateFont(this UITextField textField, IEntry entry, IFontManager fontManager)
+		{
+			var uiFont = fontManager.GetFont(entry.Font);
+			textField.Font = uiFont;
+		}
+
+		public static void UpdateReturnType(this UITextField textField, IEntry entry)
+		{
+			textField.ReturnKeyType = entry.ReturnType.ToNative();
+		}
+
+		public static void UpdateCharacterSpacing(this UITextField textField, IText textView)
+		{
+			var textAttr = textField.AttributedText?.AddCharacterSpacing(textView.Text, textView.CharacterSpacing);
 
 			if (textAttr != null)
 				textField.AttributedText = textAttr;
-
-			// TODO: Include AttributedText to Label Placeholder
 		}
 
-		public static void UpdateFont(this UITextField textField, ISearchBar searchBar, IFontManager fontManager)
+		public static void UpdateFont(this UITextField textField, IText textView, IFontManager fontManager)
 		{
-			var uiFont = fontManager.GetFont(searchBar.Font);
+			var uiFont = fontManager.GetFont(textView.Font);
 			textField.Font = uiFont;
-
-			textField.UpdateCharacterSpacing(searchBar);
 		}
 	}
 }
